tests: moved conftest around
